+++ war/vis/TestVis.js
@@ -240,6 +240,8 @@
 
             if ((Math.floor(Math.random()  * 1000) % 2) == 0) {
                 settings.isShowTagsEnabled = "True";
+            } else {
+                settings.isShowTagsEnabled = "False";
             }
 
             testData = new TestData();